tests/wpt/tests/html/canvas/offscreen/filters/2d.filter.layers.blur.exceptions.tentative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.filter.layers.blur.exceptions.tentative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.filter.layers.blur.exceptions.tentative</h1>
<p class="desc">Test exceptions on gaussianBlur filter</p>


<script>
var t = async_test("Test exceptions on gaussianBlur filter");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'gaussianBlur'}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'gaussianBlur', stdDeviation: undefined}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'gaussianBlur', stdDeviation: 'foo'}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'gaussianBlur', stdDeviation: [1,2,3]}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'gaussianBlur', stdDeviation: NaN}}); });
  assert_throws_js(TypeError, function() { ctx.beginLayer({filter:
    {name: 'gaussianBlur', stdDeviation: {}}}); });
  t.done();

});
</script>
